refactor: relocate localization structure and configure release-only Crashlytics (#6)
