+++ lib/main.dart
@@ -7,7 +7,7 @@ import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:nuntium/config/dependency_injection.dart';
 import 'package:nuntium/config/routes.dart';
 import 'package:nuntium/core/resources/app_assets.dart';
-import 'package:nuntium/core/services/language_config.dart';
+import 'package:nuntium/core/localization/language_config.dart';
 
 import 'core/theme/app_theme.dart';
 import 'core/utils/app_logger.dart';
